backend/requirements.txt
Flask
Flask-CORS
Flask-PyMongo
python-dotenv
cloudinary
flask_bcrypt

